allow to control watch activity by touching/deleting a special file
The path to the file must be given by configuration (noop_file). If that file
is deleted, the AnyEvent::Filesys::Notify Watcher is disabled and process_files
is never called. The Cron watcher re-enables the Notify watcher, when the
noop_file is touched.

To make this easier, all AnyEvent Watchers are now handled within Action::Base
instead of both Daemon.pm and Action/Base.pm.